+++ eras/mary/impl/src/Cardano/Ledger/Mary/TxBody.hs
@@ -1,6 +1,7 @@
 module Cardano.Ledger.Mary.TxBody (
   MaryEraTxBody (..),
   MaryTxBody (
+    MkMaryTxBody,
     MaryTxBody,
     mtbAuxDataHash,
     mtbCerts,