+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Tx/Internal.hs
@@ -4,7 +4,6 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -13,6 +12,7 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
+{-# LANGUAGE ViewPatterns #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 {-# OPTIONS_HADDOCK not-home #-}
 
@@ -62,17 +62,16 @@ import Cardano.Ledger.Binary.Coders
 import qualified Cardano.Ledger.Binary.Plain as Plain
 import Cardano.Ledger.Coin (Coin)
 import Cardano.Ledger.Core
-import Cardano.Ledger.Keys.Bootstrap (bootstrapWitKeyHash)
-import Cardano.Ledger.Keys.WitVKey (witVKeyHash)
 import Cardano.Ledger.MemoBytes (
   EqRaw (..),
   Mem,
   MemoBytes,
   Memoized (..),
-  memoBytesEra,
-  mkMemoBytes,
-  pattern Memo,
+  getMemoRawBytes,
+  getMemoRawType,
+  lensMemoRawType,
  )
+import Cardano.Ledger.MemoBytes.Internal (memoBytesEra, mkMemoBytes)
 import Cardano.Ledger.Shelley.Era (ShelleyEra)
 import Cardano.Ledger.Shelley.Scripts (validateMultiSig)
 import Cardano.Ledger.Shelley.TxAuxData ()
@@ -89,11 +88,10 @@ import Data.Maybe.Strict (
   strictMaybeToMaybe,
  )
 import Data.Set (Set)
-import qualified Data.Set as Set
 import Data.Typeable (Typeable)
 import Data.Word (Word32)
 import GHC.Generics (Generic)
-import Lens.Micro (Lens', SimpleGetter, lens, to, (^.))
+import Lens.Micro (Lens', SimpleGetter, to, (^.))
 import NoThunks.Class (NoThunks (..))
 
 -- ========================================================
@@ -101,7 +99,7 @@ import NoThunks.Class (NoThunks (..))
 data ShelleyTxRaw era = ShelleyTxRaw
   { strBody :: !(TxBody era)
   , strWits :: !(TxWits era)
-  , strAuxiliaryData :: !(StrictMaybe (TxAuxData era))
+  , strAuxData :: !(StrictMaybe (TxAuxData era))
   }
   deriving (Generic, Typeable)
 
@@ -136,7 +134,7 @@ instance
   ) =>
   NoThunks (ShelleyTxRaw era)
 
-newtype ShelleyTx era = TxConstr (MemoBytes (ShelleyTxRaw era))
+newtype ShelleyTx era = MkShelleyTx (MemoBytes (ShelleyTxRaw era))
   deriving newtype (SafeToHash, ToCBOR)
   deriving (Generic)
 
@@ -147,35 +145,30 @@ instance Memoized (ShelleyTx era) where
 -- memoized binary representation.
 bodyShelleyTxL :: forall era. EraTx era => Lens' (ShelleyTx era) (TxBody era)
 bodyShelleyTxL =
-  lens (\(TxConstr (Memo tx _)) -> strBody tx) $
-    \(TxConstr (Memo tx _)) txBody ->
-      TxConstr $ memoBytesEra @era $ encodeShelleyTxRaw $ tx {strBody = txBody}
+  lensMemoRawType @era strBody $ \tx txBody -> tx {strBody = txBody}
 {-# INLINEABLE bodyShelleyTxL #-}
 
 -- | `TxWits` setter and getter for `ShelleyTx`. The setter does update
 -- memoized binary representation.
 witsShelleyTxL :: forall era. EraTx era => Lens' (ShelleyTx era) (TxWits era)
 witsShelleyTxL =
-  lens (\(TxConstr (Memo tx _)) -> strWits tx) $
-    \(TxConstr (Memo tx _)) txWits ->
-      TxConstr $ memoBytesEra @era $ encodeShelleyTxRaw $ tx {strWits = txWits}
+  lensMemoRawType @era strWits $ \tx txWits -> tx {strWits = txWits}
 {-# INLINEABLE witsShelleyTxL #-}
 
 -- | `TxAuxData` setter and getter for `ShelleyTx`. The setter does update
 -- memoized binary representation.
-auxDataShelleyTxL :: EraTx era => Lens' (ShelleyTx era) (StrictMaybe (TxAuxData era))
+auxDataShelleyTxL :: forall era. EraTx era => Lens' (ShelleyTx era) (StrictMaybe (TxAuxData era))
 auxDataShelleyTxL =
-  lens (\(TxConstr (Memo tx _)) -> strAuxiliaryData tx) $
-    \(TxConstr (Memo tx _)) auxData -> mkShelleyTx $ tx {strAuxiliaryData = auxData}
+  lensMemoRawType @era strAuxData $ \tx txAuxData -> tx {strAuxData = txAuxData}
 {-# INLINEABLE auxDataShelleyTxL #-}
 
 -- | Size getter for `ShelleyTx`.
 sizeShelleyTxF :: SimpleGetter (ShelleyTx era) Integer
-sizeShelleyTxF = to (\(TxConstr (Memo _ bytes)) -> fromIntegral $ SBS.length bytes)
+sizeShelleyTxF = to (\(getMemoRawBytes -> bytes) -> fromIntegral $ SBS.length bytes)
 {-# INLINEABLE sizeShelleyTxF #-}
 
 wireSizeShelleyTxF :: SimpleGetter (ShelleyTx era) Word32
-wireSizeShelleyTxF = to $ \(TxConstr (Memo _ bytes)) ->
+wireSizeShelleyTxF = to $ \(getMemoRawBytes -> bytes) ->
   let n = SBS.length bytes
    in if n <= fromIntegral (maxBound :: Word32)
         then fromIntegral n
@@ -183,7 +176,7 @@ wireSizeShelleyTxF = to $ \(TxConstr (Memo _ bytes)) ->
 {-# INLINEABLE wireSizeShelleyTxF #-}
 
 mkShelleyTx :: forall era. EraTx era => ShelleyTxRaw era -> ShelleyTx era
-mkShelleyTx = TxConstr . memoBytesEra @era . encodeShelleyTxRaw
+mkShelleyTx = MkShelleyTx . memoBytesEra @era . encodeShelleyTxRaw
 {-# INLINEABLE mkShelleyTx #-}
 
 mkBasicShelleyTx :: EraTx era => TxBody era -> ShelleyTx era
@@ -192,7 +185,7 @@ mkBasicShelleyTx txBody =
     ShelleyTxRaw
       { strBody = txBody
       , strWits = mkBasicTxWits
-      , strAuxiliaryData = SNothing
+      , strAuxData = SNothing
       }
 
 instance EraTx ShelleyEra where
@@ -270,15 +263,13 @@ pattern ShelleyTx ::
   StrictMaybe (TxAuxData era) ->
   ShelleyTx era
 pattern ShelleyTx {body, wits, auxiliaryData} <-
-  TxConstr
-    ( Memo
-        ShelleyTxRaw
-          { strBody = body
-          , strWits = wits
-          , strAuxiliaryData = auxiliaryData
-          }
-        _
-      )
+  ( getMemoRawType ->
+      ShelleyTxRaw
+        { strBody = body
+        , strWits = wits
+        , strAuxData = auxiliaryData
+        }
+    )
   where
     ShelleyTx b w a = mkShelleyTx $ ShelleyTxRaw b w a
 
@@ -292,11 +283,11 @@ encodeShelleyTxRaw ::
   (EncCBOR (TxWits era), EncCBOR (TxBody era), EncCBOR (TxAuxData era)) =>
   ShelleyTxRaw era ->
   Encode ('Closed 'Dense) (ShelleyTxRaw era)
-encodeShelleyTxRaw ShelleyTxRaw {strBody, strWits, strAuxiliaryData} =
+encodeShelleyTxRaw ShelleyTxRaw {strBody, strWits, strAuxData} =
   Rec ShelleyTxRaw
     !> To strBody
     !> To strWits
-    !> E (encodeNullMaybe encCBOR . strictMaybeToMaybe) strAuxiliaryData
+    !> E (encodeNullMaybe encCBOR . strictMaybeToMaybe) strAuxData
 
 instance
   (Era era, EncCBOR (TxWits era), EncCBOR (TxBody era), EncCBOR (TxAuxData era)) =>
@@ -369,7 +360,7 @@ unsafeConstructTxWithBytes ::
   StrictMaybe (TxAuxData era) ->
   LBS.ByteString ->
   ShelleyTx era
-unsafeConstructTxWithBytes b w a bytes = TxConstr (mkMemoBytes (ShelleyTxRaw b w a) bytes)
+unsafeConstructTxWithBytes b w a bytes = MkShelleyTx (mkMemoBytes (ShelleyTxRaw b w a) bytes)
 
 --------------------------------------------------------------------------------
 -- Segregated witness
@@ -407,21 +398,21 @@ segWitTx ::
   TxWits era ->
   Maybe (TxAuxData era) ->
   ShelleyTx era
-segWitTx body' witnessSet metadata =
+segWitTx body' witnessSet auxData =
   let
-    wrappedMetadataBytes = case metadata of
+    wrappedAuxDataBytes = case auxData of
       Nothing -> Plain.serialize Plain.encodeNull
       Just b -> Plain.serialize b
     fullBytes =
       Plain.serialize (Plain.encodeListLen 3)
         <> Plain.serialize body'
         <> Plain.serialize witnessSet
-        <> wrappedMetadataBytes
+        <> wrappedAuxDataBytes
    in
     unsafeConstructTxWithBytes
       body'
       witnessSet
-      (maybeToStrictMaybe metadata)
+      (maybeToStrictMaybe auxData)
       fullBytes
 
 -- ========================================
@@ -436,6 +427,5 @@ witsFromTxWitnesses ::
   EraTx era =>
   Tx era ->
   Set (KeyHash 'Witness)
-witsFromTxWitnesses tx =
-  Set.map witVKeyHash (tx ^. witsTxL . addrTxWitsL)
-    `Set.union` Set.map bootstrapWitKeyHash (tx ^. witsTxL . bootAddrTxWitsL)
+witsFromTxWitnesses tx = keyHashWitnessesTxWits (tx ^. witsTxL)
+{-# DEPRECATED witsFromTxWitnesses "In favor ot `keyHashWitnessesTxWits`" #-}