+++ eras/babbage/impl/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Version history for `cardano-ledger-babbage`
 
-## 1.11.0.1
+## 1.11.1.0
+
+* Expose access to `BabbageTxBodyRaw`
+* Expose constructor `MkBabbageTxBody`
+* Deprecate `spendInputs'`, `collateralInputs'`, `referenceInputs'`, `outputs'`, `collateralReturn'`, `totalCollateral'`, `certs'`, `txfee'`, `withdrawals'`, `vldt'`, `update'`, `reqSignerHashes'`, `adHash'`, `mint'`, `scriptIntegrityHash'`, `spendInputs'`, `txnetworkid'`.
 
 
 ## 1.11.0.0