+++ eras/shelley/impl/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.17.0.0
 
+* Deprecate `witsFromTxWitnesses`
+* Expose access to `ShelleyTxRaw`, `ShelleyTxAuxDataRaw`, `ShelleyTxBodyRaw`, `ShelleyTxWitsRaw`, `MkMultiSig`
+* Expose constructors `MkShelleyTx`, `MkShelleyTxAuxData`, `MkShelleyTxBody`, `MkShelleyTxWits`, `MultiSigRaw`
+* Remove re-export of `WitVKey` from `Cardano.Ledger.Shelley.TxWits`
 * Deprecate `esAccountStateL`, `epochStateTreasuryL`, `asTreasuryL` and `asReservesL`
 * Rename `esAccountState` to `esChainAccountState`
 * Replaced `prDState` and `prPState` with `prCertState` in `ShelleyPoolreapState`